+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added `Beaker.(job|experiment).follow_structured()` and `Beaker.(job|experiment).structured_logs()` methods, which stream structured log events using the RPC interface.
+
 ## [v1.35.0](https://github.com/allenai/beaker-py/releases/tag/v1.35.0) - 2025-04-21
 
 ### Added